requirements.txt
# Core Dependencies
streamlit>=1.22.0
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.2.2
#pyarrow==12.0.0
matplotlib==3.8.0
seaborn==0.13.2
nltk>=3.8.1


# Utilities
requests>=2.28.2
typing_extensions==4.8.0
tenacity==8.2.3

# Streamlit Extensions (Only if used)
st-pages==0.4.5
streamlit-option-menu==0.3.6
streamlit-card==0.0.61

Pillow>=9.5.0
